refactor : CardDto 클래스를 삭제하고 RequestCardDto, ResponseCardDto 클래스 작성
